requirements.txt
fastapi
fastapi-cli
uvicorn
yt-dlp
google-api-python-client
hashlib
python-dotenv